llm_docify/requirements.txt
fastapi
requests
beautifulsoup4
readability-lxml
html2text
uvicorn
trafilatura